requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
pydantic>=2.7.0
python-dotenv>=1.0.1
psycopg2-binary>=2.9.9
httpx>=0.27.0
anthropic>=0.30.0
google-api-python-client>=2.137.0
google-auth>=2.31.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
